+++ src/main/java/argo/cost/common/entity/UserRole.java
@@ -22,7 +22,7 @@ public class UserRole implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy=GenerationType.TABLE)
-	private long id;
+	private Integer id;
 
 	@Column(name="created_date")
 	private Timestamp createdDate;
@@ -49,11 +49,11 @@ public class UserRole implements Serializable {
 	public UserRole() {
 	}
 
-	public long getId() {
+	public Integer getId() {
 		return this.id;
 	}
 
-	public void setId(long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
